leveldb: create the DB if it doesn't already exist.
Add a FileSystem.Stat method and change memfs' errors to satisfy
os.IsNotExist when appropriate.

Add a ErrorIfDBExists option. The C++ code also has a CreateIfMissing
option, but I feel like you'd always want to set it to true (IIUC the
C++ code still writes and possibly deletes files even if you open a
DB 'just for reading'), and so I've left it out of the Go options.

R=bradfitz
CC=golang-dev
https://codereview.appspot.com/17550044